+++ src/PSDocs.Azure/docs/Azure.Template.Doc.ps1
@@ -106,7 +106,33 @@ function global:GetTemplateMetadata {
         [String]$Path
     )
     process {
-        return (Get-Content -Path $Path -Raw | ConvertFrom-Json).metadata;
+        $template = Get-Content -Path $Path -Raw | ConvertFrom-Json;
+        if ([bool]$template.PSObject.Properties['metadata']) {
+            return $template.metadata;
+        }
+    }
+}
+
+# A function to import outputs
+function global:GetTemplateOutput {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $True)]
+        [String]$Path
+    )
+    process {
+        $template = Get-Content -Path $Path -Raw | ConvertFrom-Json;
+        foreach ($property in $template.outputs.PSObject.Properties) {
+            $output = [PSCustomObject]@{
+                Name = $property.Name
+                Type = $property.Value.type
+                Description = ''
+            }
+            if ([bool]$property.Value.PSObject.Properties['metadata'] -and [bool]$property.Value.metadata.PSObject.Properties['description']) {
+                $output.Description = $property.Value.metadata.description
+            }
+            $output;
+        }
     }
 }
 
@@ -117,16 +143,23 @@ Document 'README' {
     $templatePath = $InputObject;
     $parameters = GetTemplateParameter -Path $templatePath;
     $metadata = GetTemplateMetadata -Path $templatePath;
+    $outputs = GetTemplateOutput -Path $templatePath;
 
     # Set document title
-    Title $metadata.name
+    if ($Null -ne $metadata -and [bool]$metadata.PSObject.Properties['name']) {
+        Title $metadata.name
+    }
+    else {
+        Title $LocalizedData.DefaultTitle
+    }
 
     # Write opening line
     $metadata.Description
 
-    # Add each parameter to a table
+    # Add table and detail for each parameter
     Section $LocalizedData.Parameters {
-        $parameters | Table -Property @{ Name = $LocalizedData.ParameterName; Expression = { $_.Name }}, $LocalizedData.Description
+        $parameters | Table -Property @{ Name = $LocalizedData.ParameterName; Expression = { $_.Name }},
+            @{ Name = $LocalizedData.Description; Expression = { $_.Description }}
 
         foreach ($parameter in $parameters) {
             Section $parameter.Name {
@@ -145,6 +178,14 @@ Document 'README' {
         }
     }
 
+    # Add table for outputs
+    Section $LocalizedData.Outputs {
+        $outputs | Table -Property @{ Name = $LocalizedData.Name; Expression = { $_.Name }},
+            @{ Name = $LocalizedData.Type; Expression = { $_.Type }},
+            @{ Name = $LocalizedData.Description; Expression = { $_.Description }}
+    }
+
+    # Insert snippet
     $example = GetTemplateExample -Path $templatePath;
     Section $LocalizedData.Snippets {
         $example | Code 'json'