+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 What's changed since v0.1.0:
 
+- New features:
+  - Template outputs are added to generated document. [#28](https://github.com/Azure/PSDocs.Azure/issues/28)
 - Bug fixes:
   - Fixed snippet with short relative template causes exception. [#26](https://github.com/Azure/PSDocs.Azure/issues/26)
+  - Fixed cannot bind argument when metadata name is not set. [#35](https://github.com/Azure/PSDocs.Azure/issues/35)
 
 ## v0.1.0